Disable "Smart Quotes" in the query editor, among others.
It might be useful to keep this setting on globally, but smart quotes break
SQL, resulting in the requirement of having to do an undo-type-undo dance to
force the editor to not replace the single (') and double (") quote
characters with the ‘, ’ and “, ” characters, respectively.  Since we
can presume that no one will want to use these non-SQL quotation marks
within SQL unless they explicitly decide to do so (in matching a varchar
field, for instance), they should be off by default.  Additionally, the
double-dash (--) is used in SQL for inline comments, which gets substituted
for the em-dash (—). Disabling this seems to be a good idea on the same
argument.

More info about text substitution: http://support.apple.com/kb/PH14294