Another attempt to fix the cog menus on 10.10
Previously our cog menus have been Round Rect (borderless) NSPopupButtons with text aligned to the left of the image as that combination would allow us to make the box of the button exactly as big as the image.
However in 10.10 Apple changed the layout so that a padding is always added on the right border of a popup button (except if the menu is attached to the left).
From playing around I found the only setting to be compatible between 10.9 and 10.10 to be the button type Bevel (borderless).
This commit changes all cog buttons to Bevel type in hopes those will also look the same on 10.8, 10.7, 10.6.