Change lexer definition of high bytes.
flex does not(*) support UTF-8, therefore
alpha	[a-z_\.À-ﾟ]
has always been interpreted by flex as
alpha	[a-z_\.\xC3\x80-\xEF\xBE\x9F]

I assume this is not what was indetend and the only reason it worked, is because C3 (195),BE (190) and 9F (159) are already covered by 80-EF (128-239). Incidentally this range would also cover the whole Unicode BMP in UTF8.
This change should make it more obvious.

(*) There were some patches in 2012 and 2014 but they don't seem to have been merged.