Fix an issue on 10.7+ introduced in ce33534
This removes a IMHO useless makeFirstResponder: call. In all cases where [NSApp mainWindow] == usedSheet (< 10.7) the call right above it would already handle it. On the other hand if [NSApp mainWindow] != usedSheet (10.7+) the call would fail anyways because in that case editTextView is not even a subview of the main window.